requirements.txt
streamlit
transformers
openai
Pillow
faiss-cpu
torch